rust_xlsxwriter::chart

Trait IntoChartRange

pub trait IntoChartRange {
    // Required method
    fn new_chart_range(&self) -> ChartRange;
}

Trait to map types into an ChartRange.

The 2 most common types of range used in rust_xlsxwriter charts are:

A string with an Excel like range formula such as "Sheet1!$A$1:$A$3".

A 5 value tuple that can be used to create the range programmatically using a sheet name and zero indexed row and column values like: ("Sheet1", 0, 0, 2, 0) (this gives the same range as the previous string value).

For single cell ranges used in chart items such as chart or axis titles you can also use:

A simple string title.

A string with an Excel like cell formula such as "Sheet1!$A$1".

A 3 value tuple that can be used to create the cell range programmatically using a sheet name and zero indexed row and column values like: ("Sheet1", 0, 0) (this gives the same range as the previous string value).

Required Methods

fn new_chart_range(&self) -> ChartRange

Trait function to turn a type into ChartRange.