rust_xlsxwriter::conditional_format

Trait ConditionalFormat

pub trait ConditionalFormat {
    // Required methods
    fn validate(&self) -> Result<(), XlsxError>;
    fn rule(
        &self,
        dxf_index: Option<u32>,
        priority: u32,
        range: &str,
        guid: &str,
    ) -> String;
    fn x14_rule(&self, priority: u32, guid: &str) -> String;
    fn format_as_mut(&mut self) -> Option<&mut Format>;
    fn format_index(&self) -> Option<u32>;
    fn multi_range(&self) -> String;
    fn has_x14_extensions(&self) -> bool;
    fn has_x14_only(&self) -> bool;
    fn box_clone(&self) -> Box<dyn ConditionalFormat + Send>;
}

Trait for generic conditional format types.

Required Methods

fn validate(&self) -> Result<(), XlsxError>

Validate the conditional format.

Errors

XlsxError::ConditionalFormatError - A general error that is raised when a conditional formatting parameter is incorrect or missing.

fn rule( &self, dxf_index: Option<u32>, priority: u32, range: &str, guid: &str, ) -> String

Return the conditional format rule as an XML string.

fn x14_rule(&self, priority: u32, guid: &str) -> String

Return the extended x14 conditional format rule as an XML string.

fn format_as_mut(&mut self) -> Option<&mut Format>

Get a mutable reference to the format object in the conditional format.

fn format_index(&self) -> Option<u32>

Get the index of the format object in the conditional format.

fn multi_range(&self) -> String

Get the multi-cell range for the conditional format, if present.

fn has_x14_extensions(&self) -> bool

Check if the conditional format uses Excel 2010+ extensions.

fn has_x14_only(&self) -> bool

Check if the conditional format uses Excel 2010+ extensions only.

fn box_clone(&self) -> Box<dyn ConditionalFormat + Send>

Clone a reference into a concrete Box type.