rust_xlsxwriter::conditional_format

Trait IntoConditionalFormatValue

pub trait IntoConditionalFormatValue {
    // Required method
    fn new_value(self) -> ConditionalFormatValue;
}

Trait to map rust types into an ConditionalFormatValue value.

The IntoConditionalFormatValue trait is used to map Rust types like strings, numbers, dates, times and formulas into a generic type that can be used to replicate Excel data types used in Conditional Formats.

See ConditionalFormatValue for more information.

Required Methods

fn new_value(self) -> ConditionalFormatValue

Function to turn types into a ConditionalFormatValue enum value.