rust_xlsxwriter

Enum XlsxError

pub enum XlsxError {

    ParameterError(String),
    RowColumnLimitError,
    RowColumnOrderError,
    SheetnameCannotBeBlank(String),
    SheetnameLengthExceeded(String),
    SheetnameReused(String),
    SheetnameContainsInvalidCharacter(String),
    SheetnameStartsOrEndsWithApostrophe(String),
    MaxStringLengthExceeded,
    UnknownWorksheetNameOrIndex(String),
    MergeRangeSingleCell,
    MergeRangeOverlaps(String, String),
    MaxUrlLengthExceeded,
    UnknownUrlType(String),
    UnknownImageType,
    ImageDimensionError,
    ChartError(String),
    SparklineError(String),
    DateTimeRangeError(String),
    DateTimeParseError(String),
    TableRangeOverlaps(String, String),
    TableError(String),
    TableNameReused(String),
    ConditionalFormatError(String),
    DataValidationError(String),
    VbaNameError(String),
    CustomError(String),
    IoError(Error),
    ZipError(ZipError),

}

The XlsxError enum defines the error values for the rust_xlsxwriter library.

Variants

ParameterError(String)

A general parameter error that is raised when a parameter conflicts with an Excel limit or syntax. The nature of the error is in the error string.

RowColumnLimitError

Row or column argument exceeds Excel’s limits of 1,048,576 rows and 16,384 columns for a worksheet.

RowColumnOrderError

First row or column is greater than last row or column in a range specification, i.e., the order is reversed.

SheetnameCannotBeBlank(String)

Worksheet name cannot be blank.

SheetnameLengthExceeded(String)

Worksheet name exceeds Excel’s limit of 31 characters.

SheetnameReused(String)

Worksheet name is already in use in the workbook.

SheetnameContainsInvalidCharacter(String)

Worksheet name cannot contain any of the following invalid characters: [ ] : * ? / \

SheetnameStartsOrEndsWithApostrophe(String)

Worksheet name cannot start or end with an apostrophe.

MaxStringLengthExceeded

String exceeds Excel’s limit of 32,767 characters.

UnknownWorksheetNameOrIndex(String)

Error when trying to retrieve a worksheet reference by index or by name.

MergeRangeSingleCell

A merge range cannot be a single cell in Excel.

MergeRangeOverlaps(String, String)

The merge range overlaps a previous merge range. This is a strictly prohibited by Excel.

MaxUrlLengthExceeded

URL string exceeds Excel’s url of 2080 characters.

UnknownUrlType(String)

Unknown url type. The URL/URIs supported by Excel are http://, https://, ftp://, ftps://, mailto:, file:// and the pseudo-uri internal:.

UnknownImageType

Unknown image type. The supported image formats are PNG, JPG, GIF and BMP. See Image for details.

ImageDimensionError

Image has zero width or height, or the dimensions couldn’t be read.

ChartError(String)

A general error that is raised when a chart parameter is incorrect, or a chart is configured incorrectly.

SparklineError(String)

A general error that is raised when a sparkline parameter is incorrect, or a sparkline is configured incorrectly.

DateTimeRangeError(String)

A general error when one of the parameters supplied to a ExcelDateTime method is outside Excel’s allowable ranges.

Excel restricts dates to the range 1899-12-31 to 9999-12-31. For hours the range is generally 0-24 although larger ranges can be used to indicate durations. Minutes should be in the range 0-60 and seconds should be in the range 0.0-59.999. Excel only supports millisecond resolution.

DateTimeParseError(String)

A parsing error when trying to convert a string into an ExcelDateTime.

The allowable date/time formats supported by ExcelDateTime::parse_from_str() are:

Dates:
    yyyy-mm-dd

Times:
    hh:mm
    hh:mm:ss
    hh:mm:ss.sss

DateTimes:
    yyyy-mm-ddThh:mm:ss
    yyyy-mm-dd hh:mm:ss

The time part of DateTimes can contain optional or fractional seconds like the time examples. Timezone information is not supported by Excel and ignored in the parsing.

TableRangeOverlaps(String, String)

The table range overlaps a previous table range. This is a strictly prohibited by Excel.

TableError(String)

A general error that is raised when a table parameter is incorrect, or a table is configured incorrectly.

TableNameReused(String)

Table name is already in use in the workbook.

ConditionalFormatError(String)

A general error that is raised when a conditional format parameter is incorrect or missing.

DataValidationError(String)

A general error that is raised when a data validation parameter is incorrect or missing.

VbaNameError(String)

A general error raised when a VBA name doesn’t meet Excel’s criteria as defined by the following rules:

The name must be less than 32 characters.

The name can only contain word characters: letters, numbers and underscores.

The name must start with a letter.

The name cannot be blank.

CustomError(String)

A customizable error that can be used by third parties to raise errors or as a conversion target for other Error types.

IoError(Error)

Wrapper for a variety of std::io::Error errors such as file permissions when writing the xlsx file to disk. This can be caused by a non-existent parent directory or, commonly on Windows, if the file is already open in Excel.

ZipError(ZipError)

Wrapper for a variety of [zip::result::ZipError] errors from [zip::ZipWriter]. These relate to errors arising from creating the xlsx file zip container.