rust_xlsxwriter

Trait IntoCustomDateTimeUtc

pub trait IntoCustomDateTimeUtc {
    // Required method
    fn utc_datetime(self) -> String;
}

Trait to map user date types to an Excel UTC date.

Map a date to the Excel UTC date used in custom document properties.

This can be either a ExcelDateTime date instance or, if the chrono feature is enabled, a chrono::DateTime<Utc> instance.

Required Methods

fn utc_datetime(self) -> String

Trait method to convert a date into an Excel UTC date.