rust_xlsxwriter

Trait IntoDataValidationValue

pub trait IntoDataValidationValue {
    // Required method
    fn to_string_value(&self) -> String;
}

Trait to map rust types into data validation types

The IntoDataValidationValue trait is used to map Rust types like numbers, dates, times and formulas into a generic type that can be used to replicate Excel data types used in Data Validation.

Required Methods

fn to_string_value(&self) -> String

Function to turn types into a string value.