rust_xlsxwriter

Trait IntoExcelDateTime

pub trait IntoExcelDateTime {
    // Required method
    fn to_excel_serial_date(&self) -> f64;
}

Trait to map user date/time types to an Excel serial datetimes.

The rust_xlsxwriter library supports two ways of converting dates and times to Excel dates and times. The first is via the external Chrono library which has a comprehensive sets of types and functions for dealing with dates and times. The second is the inbuilt ExcelDateTime struct which provides a more limited set of methods and which only targets Excel specific dates and times.

In order to use Chrono with rust_xlsxwriter APIs you must enable the optional chrono feature when adding rust_xlsxwriter to your Cargo.toml.

Required Methods

fn to_excel_serial_date(&self) -> f64

Trait method to convert a date or time into an Excel serial datetime.