rust_xlsxwriter

Module utility

Utility functions for rust_xlsxwriter.

The rust_xlsxwriter library provides a number of utility functions for dealing with cell ranges, Chrono Serde serialization, and other helper method.

Examples:

use rust_xlsxwriter::{cell_range, column_number_to_name};

assert_eq!(column_number_to_name(1), "B");
assert_eq!(column_number_to_name(702), "AAA");

assert_eq!(cell_range(0, 0, 9, 0), "A1:A10");
assert_eq!(cell_range(1, 2, 8, 2), "C2:C9");
assert_eq!(cell_range(0, 0, 3, 4), "A1:E4");

Functions

cell_range

Convert zero indexed row and col cell numbers to a A1:B1 style range string.

cell_range_absolute

Convert zero indexed row and col cell numbers to an absolute $A$1:$B$1 style range string.

check_sheet_name

Check that a worksheet name is valid in Excel.

column_name_to_number

Convert a column string such as "A" to a zero indexed column reference.

column_number_to_name

Convert a zero indexed column cell reference to a string like "A".

row_col_to_cell

Convert zero indexed row and column cell numbers to a A1 style string.

row_col_to_cell_absolute

Convert zero indexed row and column cell numbers to an absolute $A$1 style range string.