rust_xlsxwriter::utility

Function check_sheet_name

pub fn check_sheet_name(name: &str) -> Result<(), XlsxError>

Check that a worksheet name is valid in Excel.

This function checks if an worksheet name is valid according to the Excel rules:

The name is less than 32 characters.

The name isn’t blank.

The name doesn’t contain any of the characters: [ ] : * ? / \.

The name doesn’t start or end with an apostrophe.

The worksheet name “History” isn’t allowed in English versions of Excel since it is a reserved name. However it is allowed in some other language versions so this function doesn’t raise it as an error. Overall it is best to avoid using it.

The rules for worksheet names in Excel are explained in the Microsoft Office documentation.

Parameters

name: The worksheet name. It must follow the Excel rules, shown above.

Errors

XlsxError::SheetnameCannotBeBlank - Worksheet name cannot be blank.

XlsxError::SheetnameLengthExceeded - Worksheet name exceeds Excel’s limit of 31 characters.

XlsxError::SheetnameContainsInvalidCharacter - Worksheet name cannot contain invalid characters: [ ] : * ? / \

XlsxError::SheetnameStartsOrEndsWithApostrophe - Worksheet name cannot start or end with an apostrophe.

Examples

The following example demonstrates testing for a valid worksheet name.

# // This code is available in examples/doc_utility_check_sheet_name.rs
#
# use rust_xlsxwriter::{utility, XlsxError};
#
fn main() -> Result<(), XlsxError> {
    // This worksheet name is valid.
    utility::check_sheet_name("2030-01-01")?;

    // This worksheet name isn't valid due to the forward slashes.
    utility::check_sheet_name("2030/01/01")?;

    Ok(())
}